Fix message in TypeError for null/undefined target
The text of the TypeError thrown by isObject mentioned "Sources", however the method is only used to check the target (in fact the message was a lie, as null / undefined sources work just fine).

The message has been changed to match the message thrown when calling Object.assign with a null or undefined target.